[standalone_compile] Dynamic shape handling
standalone_compile needs to get dynamic shape information from
somewhere. We add three new options for it:

1. from the passed-in graph (dynamic="from_graph"). This is the default.
2. ignore dynamic shapes, specialize entirely (dynamic=False)
3. from the current tracing context (dynamic="from_tracing_context")
4. from the mark_dynamic'ed info on the example inputs (not implemented).
I can see a use case for this where the user wants to partially
specialize the graph on some shape.

Maybe 1 and 3 are the same, I'm not sure. I figure more
options are better for the user.

Test Plan:

Reviewers:

Subscribers:

Tasks:

Tags:

ghstack-source-id: 6a8c8bad210af721b257413d22e3e3f1c05681df
Pull Request resolved: https://github.com/pytorch/pytorch/pull/151788